+++ frontend/components/MainPost.tsx
@@ -37,7 +37,7 @@ const MainPost = ({
       <form onSubmit={postProject}>
         <div className="nanum-bold">
           <div className="position-box">
-            <div>카테고리</div>
+            <div>{type === 1 ? '나의 포지션' : '카테고리'}</div>
             <div>
               <select
                 {...register('position', { value: data && data.position })}
